fixx

+++ src/main/java/ru/marushkai/datagathering/domain/AnalysisResult.java
@@ -35,5 +35,5 @@ public class AnalysisResult {
 
     private Boolean realExperience;
 
-    private Integer readyToProvideInfo;
+    private Double readyToProvideInfo;
 }
